* gcc.c-torture/execute/20011219-1.c (main): Make "i" a long.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@48296 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.c-torture/execute/20011219-1.c
@@ -42,7 +42,7 @@ foo (enum X x, const void *y)
 int
 main ()
 {
-  int i = 24;
+  long i = 24;
   if (foo (N, &i) != 24)
     abort ();
   exit (0);
